requirements.txt
# Core LangChain & LangGraph
langchain==0.1.20
langgraph==0.0.35
langchain-google-genai==1.0.1

# Google Gemini
google-generativeai==0.3.2

# Vector Store & Embeddings
sentence-transformers==2.2.2

# UI & Web
streamlit==1.29.0
streamlit-chat==0.1.1

# Data & Utilities
python-dotenv==1.0.0
pandas==2.1.4
geopy==2.4.1
faker==22.0.0

# Privacy & Security
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# PDF Processing (for policy docs)
pypdf==3.17.4